Move DebugAdapter to its own module.
